cod indentat :D

+++ public/javascript/app.js
@@ -5,30 +5,34 @@
         if (e.target.tagName === 'BUTTON') {
             switch (e.target.dataset.type) {
                 case 'edit':
-                    swal({
-                        title: 'Edit pet',
-                        html: createPetCard(
-                            {
-                                name: 'Bob',
-                                type: 'cat',
-                                created: Date.now(),
-                            },
-                            true
-                        ),
-                        confirmButtonText: 'Save',
-                        showCancelButton: true,
-                        cancelButtonText: 'Cancel',
-                        width: '50%'
-                    }, updatePet(e.target.dataset.id));
+                    swal(
+                        {
+                            title: 'Edit pet',
+                            html: createPetCard(
+                                {
+                                    name: 'Bob',
+                                    type: 'cat',
+                                    created: Date.now()
+                                },
+                                true
+                            ),
+                            confirmButtonText: 'Save',
+                            showCancelButton: true,
+                            cancelButtonText: 'Cancel',
+                            width: '50%'
+                        },
+                        updatePet(e.target.dataset.id)
+                    );
                     break;
                 case 'delete':
-                    swal({
-                        title: "Success",
-                        text: "Pet deleted",
-                        type: "success"
-                    },
-                    deletePet(e.target.dataset.id),
-                    e.target.parentElement.remove()
+                    swal(
+                        {
+                            title: 'Success',
+                            text: 'Pet deleted',
+                            type: 'success'
+                        },
+                        deletePet(e.target.dataset.id),
+                        e.target.parentElement.remove()
                     );
                     break;
                 default:
